code: Restructure and move functions to _helpers_usa.py
